Specify fly or ground on stats and skill Summon skill now can be specified with tags Removed tag for Sight. While searching enemy, instead of looking for targets with specific tag on Sight scriptable objects, now look for targets with specific on Skill scriptable objects.
