+++ Assets/Game/ScriptableObjects/Character/HeroMinion/MinionShoot.asset
@@ -12,11 +12,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ebacb745a1be425a85f141a5a4d0bbdb, type: 3}
   m_Name: MinionShoot
   m_EditorClassIdentifier: 
-  _cooldown: 2
-  _channelTime: 0.3
+  _flyOrGround: -1
+  skillDataAtLevel: []
+  _targetTags: []
+  hitEffect: {fileID: 0}
+  hitSound: 
   _gameObject: {fileID: 8356662070351016534, guid: c36c38d9a7faf4fc190cd694385bc986,
     type: 3}
-  _attackInfo:
-    damage: 10
-    _targetTags:
-    - Player
