+++ Assets/Game/ScriptableObjects/Character/Ninja/NinjaShoot.asset
@@ -12,8 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ebacb745a1be425a85f141a5a4d0bbdb, type: 3}
   m_Name: NinjaShoot
   m_EditorClassIdentifier: 
-  _gameObject: {fileID: 8356662070351016534, guid: c36c38d9a7faf4fc190cd694385bc986,
-    type: 3}
+  _flyOrGround: -1
   skillDataAtLevel:
   - level: 1
     damage: 3
@@ -23,3 +22,7 @@ MonoBehaviour:
     cooldown: 1.5
   _targetTags:
   - Player
+  hitEffect: {fileID: 0}
+  hitSound: 
+  _gameObject: {fileID: 8356662070351016534, guid: c36c38d9a7faf4fc190cd694385bc986,
+    type: 3}
