+++ Assets/Game/ScriptableObjects/Character/Tower/TowerSkill.asset
@@ -12,8 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ebacb745a1be425a85f141a5a4d0bbdb, type: 3}
   m_Name: TowerSkill
   m_EditorClassIdentifier: 
-  _gameObject: {fileID: 8356662070351016534, guid: c36c38d9a7faf4fc190cd694385bc986,
-    type: 3}
+  _flyOrGround: -1
   skillDataAtLevel:
   - level: 1
     damage: 5
@@ -23,3 +22,7 @@ MonoBehaviour:
     cooldown: 0.75
   _targetTags:
   - Player
+  hitEffect: {fileID: 0}
+  hitSound: 
+  _gameObject: {fileID: 8356662070351016534, guid: c36c38d9a7faf4fc190cd694385bc986,
+    type: 3}
