+++ Assets/Game/ScriptableObjects/Spawner/ThanatosSpawn.asset
@@ -12,6 +12,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 19c62d2ca45e34fe093b742d85904663, type: 3}
   m_Name: ThanatosSpawn
   m_EditorClassIdentifier: 
+  _flyOrGround: -1
+  _targetTags:
+  - Player
   _gameObject: {fileID: 5921409440072016248, guid: f0f7b51b50ce547dfb3e84bc0451acd5,
     type: 3}
   reachableDistance: 3
