+++ Assets/Game/ScriptableObjects/Spawner/TowerSpawn.asset
@@ -12,9 +12,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 19c62d2ca45e34fe093b742d85904663, type: 3}
   m_Name: TowerSpawn
   m_EditorClassIdentifier: 
-  _cooldown: 8
-  _channelTime: 0.1
+  _flyOrGround: -1
+  _targetTags:
+  - Player
   _gameObject: {fileID: 2751151623452151826, guid: b165f8e28f6c13c4ba6d1b8d68ea4f4d,
     type: 3}
-  spawnSize: 3
   reachableDistance: 3
+  minDistToTarget: 0
+  maxDistToTarget: 0
+  skillDataAtLevel: []
