+++ Assets/Game/Scripts/FlyOrGround.cs
@@ -1,0 +1,7 @@
+using System;
+
+[Flags] public enum FlyOrGround
+{
+    Fly = 1,
+    Ground = 2
+}
