+++ Assets/Game/Scripts/Skill/Skill.cs
@@ -111,5 +111,7 @@ public class Skill: ActorActionComponent
 
   }
 
+  public SkillData[] SkillDatas => skillDatas;
+
 }
